pyproject.toml
[project]
name = "openprobe_dev"
version = "0.1.0"
description = "Default template for PDM package"

dependencies = ["langchain==0.3.24", "langchain-core==0.3.56","langgraph==0.2.76","langchain-google-genai==2.1.3","google-ai-generativelanguage==0.6.18","python-dotenv>=0.19.0", "crawl4ai @ git+https://github.com/salzubi401/crawl4ai.git@main"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pdm]
distribution = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/deepsearch"]

[tool.uv]
python = "3.10"
